+++ src/main/java/net/sf/jabref/logic/importer/fetcher/ScienceDirect.java
@@ -21,7 +21,7 @@ import java.net.URL;
 import java.util.Objects;
 import java.util.Optional;
 
-import net.sf.jabref.logic.importer.FullTextFinder;
+import net.sf.jabref.logic.importer.FulltextFetcher;
 import net.sf.jabref.logic.util.DOI;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
@@ -40,11 +40,11 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 
 /**
- * FullTextFinder implementation that attempts to find a PDF URL at ScienceDirect.
+ * FulltextFetcher implementation that attempts to find a PDF URL at ScienceDirect.
  *
  * @see http://dev.elsevier.com/
  */
-public class ScienceDirect implements FullTextFinder {
+public class ScienceDirect implements FulltextFetcher {
     private static final Log LOGGER = LogFactory.getLog(ScienceDirect.class);
 
     private static final String API_URL = "http://api.elsevier.com/content/article/doi/";
